Split up the Set Command #5
